parser now parses all floats as doubles

+++ pbrt_parser/Parser.cpp
@@ -64,7 +64,7 @@ namespace pbrt_parser {
     TokenHandle token = next();
     if (!token)
       throw std::runtime_error("unexpected end of file\n@"+std::string(__PRETTY_FUNCTION__));
-    return std::stof(token);
+    return std::stod(token);
   }
 
   inline vec3f Parser::parseVec3f()
@@ -85,28 +85,28 @@ namespace pbrt_parser {
 
     assert(open == "[");
     affine3f xfm;
-    xfm.l.vx.x = std::stof(next()->c_str());
-    xfm.l.vx.y = std::stof(next()->c_str());
-    xfm.l.vx.z = std::stof(next()->c_str());
-    float vx_w = std::stof(next()->c_str());
+    xfm.l.vx.x = std::stod(next()->c_str());
+    xfm.l.vx.y = std::stod(next()->c_str());
+    xfm.l.vx.z = std::stod(next()->c_str());
+    float vx_w = std::stod(next()->c_str());
     assert(vx_w == 0.f);
 
-    xfm.l.vy.x = std::stof(next()->c_str());
-    xfm.l.vy.y = std::stof(next()->c_str());
-    xfm.l.vy.z = std::stof(next()->c_str());
-    float vy_w = std::stof(next()->c_str());
+    xfm.l.vy.x = std::stod(next()->c_str());
+    xfm.l.vy.y = std::stod(next()->c_str());
+    xfm.l.vy.z = std::stod(next()->c_str());
+    float vy_w = std::stod(next()->c_str());
     assert(vy_w == 0.f);
 
-    xfm.l.vz.x = std::stof(next()->c_str());
-    xfm.l.vz.y = std::stof(next()->c_str());
-    xfm.l.vz.z = std::stof(next()->c_str());
-    float vz_w = std::stof(next()->c_str());
+    xfm.l.vz.x = std::stod(next()->c_str());
+    xfm.l.vz.y = std::stod(next()->c_str());
+    xfm.l.vz.z = std::stod(next()->c_str());
+    float vz_w = std::stod(next()->c_str());
     assert(vz_w == 0.f);
 
-    xfm.p.x    = std::stof(next()->c_str());
-    xfm.p.y    = std::stof(next()->c_str());
-    xfm.p.z    = std::stof(next()->c_str());
-    float p_w  = std::stof(next()->c_str());
+    xfm.p.x    = std::stod(next()->c_str());
+    xfm.p.y    = std::stod(next()->c_str());
+    xfm.p.z    = std::stod(next()->c_str());
+    float p_w  = std::stod(next()->c_str());
     assert(p_w == 1.f);
 
     const std::string close = *next();
@@ -653,7 +653,7 @@ namespace pbrt_parser {
         next(); // '['
         float mat[16];
         for (int i=0;i<16;i++)
-          mat[i] = std::stof(next());
+          mat[i] = std::stod(next());
 
         affine3f xfm;
         xfm.l.vx = vec3f(mat[0],mat[1],mat[2]);
